requirements.txt
python-docx
python-pptx
openpyxl
googletrans==4.0.0-rc1
serbian-text-converter
PyQt6